NONE Remove global unused mock check
This was causing flakiness in our integration tests that expect an error
from a Promise.all call. If multiple endpoints inside a Promise.all call
are mocked, there's no guarantee that all of the mocks will be hit if
one of the Promises rejects, since Promise.all immediately halts
execution on a single rejection.