MDTZ-973 Rework webhook passcode generation
Added a 'webhookPasscode' column to the 'FigmaTeam' table. When a team
is configured, we generate a new UUID for the passcode and save this
against the FigmaTeam. There is a new 'validateWebhookPayloadUseCase'
that is used to validate the event type and passcode from a webhook
event before handling the event itself.