TGF-94 Add lastupdatedby and iconurl to atlassiandesign type (#229)
TGF-94 Add lastupdatedby and iconurl to atlassiandesign type
Currently the figma-for-jira integration submits designs to Atlassian when an associated file is updated or during a backfill operation. However, the submitted payload does not include information about the user that made the change to the file.
This PR introduces two fields to the payload that is submitted to Atlassian.

lastUpdatedBy: This field includes details about the user that made the change to a file. This is only populated when handling a FILE_UPDATE event. During other experiences (i.e. backfill, etc ...) the lastUpdatedBy field will remain undefined as that information is not available.
iconUrl: As the designs are ingested into a generic design schema, this field will have a url to the resource icon of the design submitted. As it stands, there is no resource specific icon for Figma, it will remain unset.